requirements.txt
transformers==4.37.2
datasets==2.16.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
tqdm==4.66.1
pyyaml==6.0.1
wandb==0.16.1
python-dotenv==1.0.0
sentencepiece==0.1.99
tokenizers==0.13.3 